MGDOBR-225: renaming
