+++ rhoas-client/src/test/java/com/redhat/service/bridge/rhoas/RhoasInjectionTest.java
@@ -1,0 +1,29 @@
+package com.redhat.service.bridge.rhoas;
+
+import javax.enterprise.inject.Instance;
+import javax.inject.Inject;
+
+import org.junit.jupiter.api.Test;
+
+import com.redhat.service.bridge.test.rhoas.testprofiles.RhoasTestProfile;
+
+import io.quarkus.test.junit.QuarkusTest;
+import io.quarkus.test.junit.TestProfile;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatNoException;
+
+@QuarkusTest
+@TestProfile(RhoasTestProfile.class)
+class RhoasInjectionTest {
+
+    @Inject
+    Instance<RhoasClient> rhoasClient;
+
+    @Test
+    void test() {
+        assertThat(rhoasClient.isUnsatisfied()).isFalse();
+        assertThat(rhoasClient.isAmbiguous()).isFalse();
+        assertThatNoException().isThrownBy(() -> rhoasClient.get().toString());
+    }
+}
