+++ rhoas-client/src/test/java/com/redhat/service/bridge/rhoas/RhoasInjectionWithMissingInstanceApiPropsTest.java
@@ -1,21 +1,25 @@
 package com.redhat.service.bridge.rhoas;
 
+import java.util.NoSuchElementException;
+
 import javax.enterprise.inject.Instance;
 import javax.inject.Inject;
 
 import org.junit.jupiter.api.Test;
 
-import com.redhat.service.bridge.test.rhoas.testprofiles.RhoasEnabledTestProfile;
+import com.redhat.service.bridge.test.rhoas.testprofiles.RhoasWithMissingInstanceApiPropsTestProfile;
 
 import io.quarkus.test.junit.QuarkusTest;
 import io.quarkus.test.junit.TestProfile;
 
+import static com.redhat.service.bridge.rhoas.RhoasProperties.INSTANCE_API_HOST;
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
 import static org.assertj.core.api.Assertions.assertThatNoException;
 
 @QuarkusTest
-@TestProfile(RhoasEnabledTestProfile.class)
-class RhoasInjectionEnabledTest {
+@TestProfile(RhoasWithMissingInstanceApiPropsTestProfile.class)
+class RhoasInjectionWithMissingInstanceApiPropsTest {
 
     @Inject
     Instance<RhoasClient> rhoasClient;
@@ -25,5 +29,8 @@ class RhoasInjectionEnabledTest {
         assertThat(rhoasClient.isUnsatisfied()).isFalse();
         assertThat(rhoasClient.isAmbiguous()).isFalse();
         assertThatNoException().isThrownBy(() -> rhoasClient.get().toString());
+        assertThatExceptionOfType(NoSuchElementException.class)
+                .isThrownBy(() -> rhoasClient.get().deleteTopic("foo-topic"))
+                .withMessageContaining(INSTANCE_API_HOST);
     }
 }
