+++ rhoas-client/src/test/java/com/redhat/service/bridge/rhoas/RhoasInjectionWithMissingMgmtApiPropsTest.java
@@ -7,19 +7,20 @@ import javax.inject.Inject;
 
 import org.junit.jupiter.api.Test;
 
-import com.redhat.service.bridge.test.rhoas.testprofiles.RhoasEnabledWithMissingInstanceApiPropsTestProfile;
+import com.openshift.cloud.api.kas.models.ServiceAccountRequest;
+import com.redhat.service.bridge.test.rhoas.testprofiles.RhoasWithMissingMgmtApiPropsTestProfile;
 
 import io.quarkus.test.junit.QuarkusTest;
 import io.quarkus.test.junit.TestProfile;
 
-import static com.redhat.service.bridge.rhoas.RhoasProperties.INSTANCE_API_HOST;
+import static com.redhat.service.bridge.rhoas.RhoasProperties.MGMT_API_HOST;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
 import static org.assertj.core.api.Assertions.assertThatNoException;
 
 @QuarkusTest
-@TestProfile(RhoasEnabledWithMissingInstanceApiPropsTestProfile.class)
-class RhoasInjectionEnabledWithMissingInstanceApiPropsTest {
+@TestProfile(RhoasWithMissingMgmtApiPropsTestProfile.class)
+class RhoasInjectionWithMissingMgmtApiPropsTest {
 
     @Inject
     Instance<RhoasClient> rhoasClient;
@@ -30,7 +31,7 @@ class RhoasInjectionEnabledWithMissingInstanceApiPropsTest {
         assertThat(rhoasClient.isAmbiguous()).isFalse();
         assertThatNoException().isThrownBy(() -> rhoasClient.get().toString());
         assertThatExceptionOfType(NoSuchElementException.class)
-                .isThrownBy(() -> rhoasClient.get().deleteTopic("foo-topic"))
-                .withMessageContaining(INSTANCE_API_HOST);
+                .isThrownBy(() -> rhoasClient.get().createServiceAccount(new ServiceAccountRequest().name("foo").description("bar")))
+                .withMessageContaining(MGMT_API_HOST);
     }
 }
