+++ test-utils/src/main/java/com/redhat/service/bridge/test/rhoas/testprofiles/RhoasTestProfile.java
@@ -5,7 +5,7 @@ import java.util.Map;
 
 import io.quarkus.test.junit.QuarkusTestProfile;
 
-public class RhoasEnabledWithMissingInstanceApiPropsTestProfile implements QuarkusTestProfile {
+public class RhoasTestProfile implements QuarkusTestProfile {
 
     @Override
     public Map<String, String> getConfigOverrides() {
@@ -17,6 +17,12 @@ public class RhoasEnabledWithMissingInstanceApiPropsTestProfile implements Quark
         config.put("event-bridge.rhoas.sso.red-hat.client-id", "${rhoas-mock-server.sso.red-hat.client-id}");
         config.put("event-bridge.rhoas.sso.red-hat.refresh-token", "${rhoas-mock-server.sso.red-hat.refresh-token}");
 
+        config.put("event-bridge.rhoas.instance-api.host", "${rhoas-mock-server.url:}${rhoas-mock-server.instance-api.base-path}/rest");
+
+        config.put("event-bridge.rhoas.sso.mas.auth-server-url", "${rhoas-mock-server.url:}${rhoas-mock-server.sso.mas.base-path}");
+        config.put("event-bridge.rhoas.sso.mas.client-id", "${rhoas-mock-server.sso.mas.client-id}");
+        config.put("event-bridge.rhoas.sso.mas.client-secret", "${rhoas-mock-server.sso.mas.client-secret}");
+
         config.put("quarkus.oidc-client.red-hat-sso.discovery-enabled", "false");
         config.put("quarkus.oidc-client.red-hat-sso.token-path", "/token");
         config.put("quarkus.oidc-client.mas-sso.discovery-enabled", "false");
@@ -24,5 +30,4 @@ public class RhoasEnabledWithMissingInstanceApiPropsTestProfile implements Quark
 
         return config;
     }
-
 }
