+++ test-utils/src/main/java/com/redhat/service/bridge/test/rhoas/testprofiles/RhoasWithMissingInstanceApiPropsTestProfile.java
@@ -5,17 +5,17 @@ import java.util.Map;
 
 import io.quarkus.test.junit.QuarkusTestProfile;
 
-public class RhoasEnabledWithMissingMgmtApiPropsTestProfile implements QuarkusTestProfile {
+public class RhoasWithMissingInstanceApiPropsTestProfile implements QuarkusTestProfile {
 
     @Override
     public Map<String, String> getConfigOverrides() {
         Map<String, String> config = new HashMap<>();
 
-        config.put("event-bridge.rhoas.instance-api.host", "${rhoas-mock-server.url:}${rhoas-mock-server.instance-api.base-path}/rest");
+        config.put("event-bridge.rhoas.mgmt-api.host", "${rhoas-mock-server.url:}${rhoas-mock-server.mgmt-api.base-path}");
 
-        config.put("event-bridge.rhoas.sso.mas.auth-server-url", "${rhoas-mock-server.url:}${rhoas-mock-server.sso.mas.base-path}");
-        config.put("event-bridge.rhoas.sso.mas.client-id", "${rhoas-mock-server.sso.mas.client-id}");
-        config.put("event-bridge.rhoas.sso.mas.client-secret", "${rhoas-mock-server.sso.mas.client-secret}");
+        config.put("event-bridge.rhoas.sso.red-hat.auth-server-url", "${rhoas-mock-server.url:}${rhoas-mock-server.sso.red-hat.base-path}");
+        config.put("event-bridge.rhoas.sso.red-hat.client-id", "${rhoas-mock-server.sso.red-hat.client-id}");
+        config.put("event-bridge.rhoas.sso.red-hat.refresh-token", "${rhoas-mock-server.sso.red-hat.refresh-token}");
 
         config.put("quarkus.oidc-client.red-hat-sso.discovery-enabled", "false");
         config.put("quarkus.oidc-client.red-hat-sso.token-path", "/token");
@@ -24,4 +24,5 @@ public class RhoasEnabledWithMissingMgmtApiPropsTestProfile implements QuarkusTe
 
         return config;
     }
+
 }
