+++ test-utils/src/main/java/com/redhat/service/bridge/test/rhoas/testprofiles/RhoasWithMissingMgmtApiPropsTestProfile.java
@@ -5,17 +5,11 @@ import java.util.Map;
 
 import io.quarkus.test.junit.QuarkusTestProfile;
 
-public class RhoasEnabledTestProfile implements QuarkusTestProfile {
+public class RhoasWithMissingMgmtApiPropsTestProfile implements QuarkusTestProfile {
 
     @Override
     public Map<String, String> getConfigOverrides() {
         Map<String, String> config = new HashMap<>();
-
-        config.put("event-bridge.rhoas.mgmt-api.host", "${rhoas-mock-server.url:}${rhoas-mock-server.mgmt-api.base-path}");
-
-        config.put("event-bridge.rhoas.sso.red-hat.auth-server-url", "${rhoas-mock-server.url:}${rhoas-mock-server.sso.red-hat.base-path}");
-        config.put("event-bridge.rhoas.sso.red-hat.client-id", "${rhoas-mock-server.sso.red-hat.client-id}");
-        config.put("event-bridge.rhoas.sso.red-hat.refresh-token", "${rhoas-mock-server.sso.red-hat.refresh-token}");
 
         config.put("event-bridge.rhoas.instance-api.host", "${rhoas-mock-server.url:}${rhoas-mock-server.instance-api.base-path}/rest");
